Updated database to support bigger telegram IDs

+++ asset/sql/structure.sql
@@ -3,7 +3,7 @@ USE `better_location`;
 
 CREATE TABLE `better_location_user` (
   `user_id` int(11) NOT NULL AUTO_INCREMENT,
-  `user_telegram_id` int(11) NOT NULL,
+  `user_telegram_id` bigint NOT NULL,
   `user_telegram_name` varchar(255) COLLATE utf8mb4_bin DEFAULT NULL,
   `user_registered` datetime NOT NULL DEFAULT CURRENT_TIMESTAMP,
   `user_last_update` datetime NOT NULL,
@@ -16,7 +16,7 @@ CREATE TABLE `better_location_user` (
 
 CREATE TABLE `better_location_chat` (
   `chat_id` int(11) NOT NULL AUTO_INCREMENT,
-  `chat_telegram_id` bigint(8) NOT NULL,
+  `chat_telegram_id` bigint NOT NULL,
   `chat_telegram_type` varchar(20) COLLATE utf8mb4_bin NOT NULL,
   `chat_telegram_name` varchar(255) COLLATE utf8mb4_bin DEFAULT NULL,
   `chat_registered` datetime NOT NULL DEFAULT CURRENT_TIMESTAMP,
@@ -78,7 +78,7 @@ CREATE TABLE IF NOT EXISTS better_location_web_login
     id               int auto_increment
         primary key,
     hash             varchar(64)  null,
-    user_telegram_id int          not null,
+    user_telegram_id bigint       not null,
     auth_date        int          not null,
     user_first_name  varchar(255) not null,
     user_last_name   varchar(255) null,
